refactor: remove commented code and empty files
