+++ app/[locale]/(protected)/_components/body-editor/index.ts
@@ -1,4 +1,3 @@
 export { BodyEditor } from "./body-editor";
 export { BodyEditorToolbar } from "./body-editor-toolbar";
-export { useBodyEditor } from "./use-body-editor";
 export { isJsonLike } from "./utils";
