--- entities/auth/library.ts
@@ -1,1 +1,0 @@
-// validatePassword
