--- store/middleware/persist-variables.ts
@@ -1,1 +1,0 @@
-// this is a middleware for Redux watches for changes in the variablesSlice and synchronizes them with localStorage
